Registracijos langa sutvarkiau pagal MVP
